restructure project files and points to pyproject.toml of flower
