+++ src/syft_flwr/utils.py
@@ -1,4 +1,7 @@
 import hashlib
+from pathlib import Path
+
+import tomllib  # TODO: replace with tomli, as it is not supported older python versions.
 
 
 def string_to_hash_int(input_string: str) -> int:
@@ -7,3 +10,15 @@ def string_to_hash_int(input_string: str) -> int:
     hash_hex = hash_object.hexdigest()
     hash_int = int(hash_hex, 16) % (2**32)
     return hash_int
+
+
+def read_toml_file(file_path):
+    """Read a TOML file and return the data."""
+    file_path = to_path(file_path)
+    with open(file_path, "rb") as file:
+        data = tomllib.load(file)
+    return data
+
+
+def to_path(path: str) -> Path:
+    return Path(path).expanduser().resolve()
